Add the blue header bar at the top of a HTML pre-release

Add a blue header bar to the top of a HTML pre-release containing the stats release details to mimic the appearance of gov.uk This should be done for the pre-release only, and not for the final statistics publication

Usage

add_blue_header(
  header_text,
  stats_status = "National Statistics",
  pre_release_note = "Pre-release"
)

Arguments

header_text

a string, containing the title of the pre-release to include in the header text

stats_status

a string, designating the statistics status of the pre-release. Defaults to "National Statistics", but can be Official or Experimental statistics

pre_release_note

a string, containing a pre-release note that will appear underneath the title in the header text. Defaults to "Pre-release".

Value

A plain text string which renders as a HTML banner

Examples

if (FALSE) { # \dontrun{
add_blue_header("Annual Road Traffic Statistics", stats_status = "Official Statistics")
} # }